[fix] moved main package to scripts folder
