+++ scripts/tiny.go
@@ -1,4 +1,4 @@
-package main
+package scripts
 
 import (
 	"fmt"
